requirements.txt
gspread
oauth2client
google-cloud-bigquery
pandas
pandas_gbq